feat: move runner functionality
